Refactor MobileMenu: hoist state so menu can closer drawer when login modal is opened
