+++ src/components/Menu/MobileMenu.tsx
@@ -1,49 +1,21 @@
-import { useState } from 'react'
-import { Button, Nav, Offcanvas } from 'react-bootstrap'
-import { useLocation } from 'react-router-dom'
-import User from '../../types/User'
-import { useEffect } from 'react'
-import useWindowSize from 'src/hooks/useWindowSize'
+import { Nav, Offcanvas, OffcanvasProps } from 'react-bootstrap'
 
-interface IMobileMenu {
-  user?: User
+interface IMobileMenu extends OffcanvasProps {
   children: React.ReactNode
 }
 
-export default function MobileMenu({ children }: IMobileMenu) {
-  const size = useWindowSize()
-  const location = useLocation()
-  const [showDrawer, setShowDrawer] = useState(true)
-  const open = () => setShowDrawer(true)
-  const close = () => setShowDrawer(false)
-
-  useEffect(() => {
-    if (location) {
-      close()
-    }
-  }, [location, size])
-
+export default function MobileMenu({ children, ...drawerProps }: IMobileMenu) {
   return (
-    <div className="d-md-none">
-      <Button variant="link" onClick={open}>
-        <i className="bi bi-list text-body fs-2" />
-      </Button>
-      <Offcanvas
-        show={showDrawer}
-        onHide={close}
-        placement="end"
-        style={{ width: 250 }}
-      >
-        <Offcanvas.Header closeButton />
-        <Offcanvas.Body>
-          <Nav
-            defaultActiveKey="/"
-            className="flex-column align-items-center fs-3"
-          >
-            {children}
-          </Nav>
-        </Offcanvas.Body>
-      </Offcanvas>
-    </div>
+    <Offcanvas placement="end" style={{ width: 250 }} {...drawerProps}>
+      <Offcanvas.Header closeButton />
+      <Offcanvas.Body>
+        <Nav
+          defaultActiveKey="/"
+          className="flex-column align-items-center fs-3"
+        >
+          {children}
+        </Nav>
+      </Offcanvas.Body>
+    </Offcanvas>
   )
 }
